DormsPage, Comments, Cleanup
Added comments to the App.css file and deleted some unused code (mostly my own dw).

There was also some stuff erroring in the background which was taken care of.

Created Dorms.css  and DormsPage.jsx, tried to add a working link in the main App page to Dorms but failed :(